move files to types dir and ignore types from coverage
