+++ packages/steveo/package.json
@@ -9,7 +9,7 @@
   "repository": "git@github.com:ordermentum/steveo.git",
   "scripts": {
     "test": "NODE_ENV=test nyc yarn spec",
-    "spec": "tsc && yarn mocha --require ts-node/register --bail=false 'test/**/*_test.ts'",
+    "spec": "tsc && yarn mocha",
     "lint": "yarn eslint 'src/**/*.{ts,js}'",
     "lint:test": "yarn eslint 'test/**/*.{ts,js}'",
     "autotest": "yarn run mocha --watch",
@@ -28,7 +28,8 @@
       "coverage",
       "examples",
       "test",
-      "lib"
+      "lib",
+      "src/types"
     ],
     "reporter": [
       "text",
